requirements.txt
# Core Python dependencies
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0

# LLM / HTTP (if API fallback is needed)
requests==2.31.0

# Optional: Gmail integration
google-api-python-client==2.115.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0

# Logging and dev
tqdm==4.66.1
